+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 - Added `pybricks.pupdevices.MarioHub` to control it as a peripheral. It
   cannot be used as a standalone device since it cannot ne updated.
+- Added `DriveBase.hold()` method ([support#2621]).
 
 ### Changed
 - The EV3 Color Sensor now returns `Color.NONE` instead of `None` when no color